+++ src/index.html
@@ -23,13 +23,14 @@
             placeholder="Type new todo"
             autofocus
             type="text"
+            id="new-todo"
           />
         </div>
       </header>
       <div class="container todoapp-wrapper">
         <!-- This section should be hidden by default and shown when there are todos -->
         <section class="main" id="main">
-          <ul class="todo-list">
+          <ul class="todo-list" id="todo-list">
             <!-- These are here just to show the structure of the list items -->
             <!-- List items should get the class `editing` when editing and `completed` when marked as completed -->
             <!-- <li class="completed">
