+++ src/index.js
@@ -1,17 +1,29 @@
 import "./css/base.css";
+import { validarListaTareas } from "./js/utils";
 
 // import { sayHello } from "./js/utils";
 //#1 Ocultar las secciones main y footer
-const main = document.querySelector("#main");
-const footer = document.querySelector("#footer");
+validarListaTareas();
+//2 agregar tarea nueva y dejarla en pendiente.
+const inputTarea = document.querySelector("#new-todo");
+const listaTareas = document.querySelector("#todo-list");
 
-const contenedorTareas = document.getElementsByClassName("todo-list");
-const tareas = contenedorTareas[0].children;
-
-if (tareas.length == 0) {
-  main.classList.add("inactive");
-  footer.classList.add("inactive");
-} else {
-  main.classList.remove("inactive");
-  footer.classList.remove("inactive");
-}
+document.addEventListener("keyup", (event) => {
+  if (event.key === "Enter") {
+    let inputNuevaTarea = inputTarea.value;
+    inputNuevaTarea = inputNuevaTarea.trim();
+    inputTarea.value = "";
+    if (inputNuevaTarea.length > 0) {
+      const htmlNuevoElemento = `<li>
+      <div class="view">
+        <input class="toggle" type="checkbox" />
+        <label>${inputNuevaTarea}</label>
+        <button class="destroy"></button>
+      </div>
+      <input class="edit" value="${inputNuevaTarea}" />
+    </li>`;
+      listaTareas.innerHTML = listaTareas.innerHTML + htmlNuevoElemento;
+      validarListaTareas();
+    }
+  }
+});
